Move 'key._KeyIterator' -> 'bucket._KeyIterator'.
The iterator is an implementation detail of 'Bucket'.
